backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.10.6
pydantic-settings>=2.3.0

# AWS / Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Strands Agents SDK
strands-agents>=1.0.0
strands-agents-tools>=0.2.0

# Nova Act
nova-act>=3.0.0

# Vector store
faiss-cpu>=1.8.0
numpy>=1.26.0

# Async / streaming
anyio>=4.4.0
httpx>=0.27.0
sse-starlette>=2.1.0

# Document processing
pypdf>=4.2.0
pillow>=11.0.0

# Utilities
python-multipart>=0.0.9
